+++ metadata.rb
@@ -1,13 +1,13 @@
-name 'wkhtmltopdf'
-maintainer 'Brian Flad'
+name             'wkhtmltopdf'
+maintainer       'Brian Flad'
 maintainer_email 'bflad417@gmail.com'
-license 'Apache 2.0'
-description 'Installs wkhtmltopdf static binaries'
-version '0.1.0'
+license          'Apache 2.0'
+description      'Installs wkhtmltoimage and wkhtmltopdf'
+long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version          '0.2.0'
 
-recipe 'wkhtmltopdf', 'Installs wkhtmltoimage and wkhtmltopdf static binaries'
-recipe 'wkhtmltopdf::wkhtmltoimage_binary', 'Installs wkhtmltoimage static binary'
-recipe 'wkhtmltopdf::wkhtmltopdf_binary', 'Installs wkhtmltopdf static binary'
+recipe 'wkhtmltopdf', 'Installs wkhtmltoimage and wkhtmltopdf'
+recipe 'wkhtmltopdf::binary', 'Installs wkhtmltoimage and wkhtmltopdf binaries'
 
 supports 'amazon'
 supports 'centos'
@@ -15,3 +15,5 @@ supports 'debian'
 supports 'fedora'
 supports 'redhat'
 supports 'ubuntu'
+
+depends 'apt'