+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.2.0
+
+* Bugfix: Removed extraneous wkhtmltoimage/wkhtmltopdf attributes
+* Enhancement: [#2][]: Updated to wkhtmltopdf 0.12.0
+* Enhancement: Consolidated _binary recipes into single binary recipe
+* Enhancement: Set lib_dir to install wkhtmltox.so
+* Enhancement: Move dependency package installation to dependency_packages attribute and default recipe
+* Enhancement: Renamed arch attribute to platform (to better handle download URLs)
+
 ## 0.1.0
 
 * Initial release
+
+[#2]: https://github.com/bflad/chef-wkhtmltopdf/issues/2